remove IssuState.all
