+++ Sources/GitHubKit/Data/Issue/IssueState.swift
@@ -7,7 +7,6 @@ import Foundation
 public enum IssueState: String, Codable, Sendable {
   case open
   case closed
-  case all
   
   public init(from decoder: Decoder) throws {
     let container = try decoder.singleValueContainer()
