+++ Sources/GitHubKit/Issue/IssueSearchSortType.swift
@@ -1,0 +1,11 @@
+//
+//  IssueSearchSortType.swift
+//
+
+import Foundation
+
+public enum IssueSearchSortType: String {
+  case created
+  case updated
+  case comments
+}
