+++ Sources/GitHubKit/Issue/IssueSearchState.swift
@@ -1,0 +1,11 @@
+//
+//  IssueSearchState.swift
+//
+
+import Foundation
+
+public enum IssueSearchState: String {
+  case open
+  case closed
+  case all
+}
